[Chartjs][Notify][React][Svelte] Remove StimulusHelper deprecation for 3.0
