+++ src/Chartjs/src/Twig/ChartExtension.php
@@ -13,7 +13,6 @@ namespace Symfony\UX\Chartjs\Twig;
 
 use Symfony\UX\Chartjs\Model\Chart;
 use Symfony\UX\StimulusBundle\Helper\StimulusHelper;
-use Symfony\WebpackEncoreBundle\Twig\StimulusTwigExtension;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFunction;
 
@@ -24,19 +23,8 @@ use Twig\TwigFunction;
  */
 class ChartExtension extends AbstractExtension
 {
-    private $stimulus;
-
-    /**
-     * @param $stimulus StimulusHelper
-     */
-    public function __construct(StimulusHelper|StimulusTwigExtension $stimulus)
+    public function __construct(private StimulusHelper $stimulusHelper)
     {
-        if ($stimulus instanceof StimulusTwigExtension) {
-            trigger_deprecation('symfony/ux-chartjs', '2.9', 'Passing an instance of "%s" to "%s" is deprecated, pass an instance of "%s" instead.', StimulusTwigExtension::class, __CLASS__, StimulusHelper::class);
-            $stimulus = new StimulusHelper(null);
-        }
-
-        $this->stimulus = $stimulus;
     }
 
     public function getFunctions(): array
@@ -56,7 +44,7 @@ class ChartExtension extends AbstractExtension
         }
         $controllers['@symfony/ux-chartjs/chart'] = ['view' => $chart->createView()];
 
-        $stimulusAttributes = $this->stimulus->createStimulusAttributes();
+        $stimulusAttributes = $this->stimulusHelper->createStimulusAttributes();
         foreach ($controllers as $name => $controllerValues) {
             $stimulusAttributes->addController($name, $controllerValues);
         }
